airflow/requirements.txt
apache-airflow-providers-google
google-cloud-storage
pyarrow
nba_api
pandas
asyncio